Tests: * Create PrimeNumberToolsReduceByPrimeRangeTest.kt - add relevant test cases from PrimeNumberToolsTest
