+++ numbers/src/test/kotlin/mathtools/numbers/primes/PrimeNumberToolsTest.kt
@@ -1,7 +1,6 @@
 package mathtools.numbers.primes
 
 import mathtools.numbers.primes.PrimeNumberTools.divideOutSmallPrimes
-import mathtools.numbers.primes.PrimeNumberTools.reduceByPrimeRange
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertFalse
 import org.junit.jupiter.api.Assertions.assertNull
@@ -84,72 +83,6 @@ class PrimeNumberToolsTest {
 			PrimeFactoring.firstPrimeAbove(0, 5, cache))
 		assertNull(
 			PrimeFactoring.firstPrimeAbove(1, 5, cache))
-	}
-
-	@ParameterizedTest
-	@ArgumentsSource(PrimeCacheArgumentProvider::class)
-	fun testReduceByPrimes(
-		cache: PrimeCacheBase
-	) {
-		assertEquals(
-			2, reduceByPrimeRange(1..2, 30, cache))
-		assertEquals(
-			4, reduceByPrimeRange(1..5, 60, cache))
-		assertEquals(
-			12, reduceByPrimeRange(2..2, 60, cache))
-	}
-
-	@ParameterizedTest
-	@ArgumentsSource(PrimeCacheArgumentProvider::class)
-	fun testReduceByPrimesNegative(
-		cache: PrimeCacheBase
-	) {
-		assertEquals(
-			-2, reduceByPrimeRange(1..2, -30, cache))
-		assertEquals(
-			-4, reduceByPrimeRange(1..2, -60, cache))
-		assertEquals(
-			-12, reduceByPrimeRange(2..3, -60, cache))
-	}
-
-	@ParameterizedTest
-	@ArgumentsSource(PrimeCacheArgumentProvider::class)
-	fun testReduceByPrimesIndivisibleProduct(
-		cache: PrimeCacheBase
-	) {
-		assertNull(
-			reduceByPrimeRange(1..3, -1, cache))
-		assertNull(
-			reduceByPrimeRange(1..3, 0, cache))
-		assertNull(
-			reduceByPrimeRange(1..3, 1, cache))
-	}
-
-	@ParameterizedTest
-	@ArgumentsSource(PrimeCacheArgumentProvider::class)
-	fun testReduceByPrimesInvalidRange(
-		cache: PrimeCacheBase
-	) {
-		// Negative valued ranges are ignored
-		assertNull(
-			reduceByPrimeRange(-3..-1, -30, cache))
-		// The invalid region of the range is ignored
-		assertNull(
-			reduceByPrimeRange(-3..3, -60, cache))
-		// Invalid region is ignored
-		assertEquals(
-			-15, reduceByPrimeRange(-1..0, -60, cache))
-	}
-
-	@ParameterizedTest
-	@ArgumentsSource(PrimeCacheArgumentProvider::class)
-	fun testReduceByPrimesReversedRange(
-		cache: PrimeCacheBase
-	) {
-		assertEquals(
-			4, reduceByPrimeRange(2..1, 60, cache))
-		assertEquals(
-			-4, reduceByPrimeRange(2..1, -60, cache))
 	}
 
 	@ParameterizedTest
